+++ migrations/000010_tesla_public_key.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS tesla_public_key;